+++ src/parser.ts
@@ -94,11 +94,11 @@ export class Parser<out T = unknown> {
             return this.run(state, context) ?? parser.run(state, context);
         });
     }
-    option(this: Parser<T>): Parser<T | null>;
+    option(this: Parser<T>): Parser<T | undefined>;
     option<U>(this: Parser<T>, value: U): Parser<T | U>;
-    option<U = null>(this: Parser<T>, value: U = null as unknown as U): Parser<T | U> {
+    option<U>(this: Parser<T>, value?: U): Parser<T | U> {
         return new Parser<T | U>((state, context) => {
-            return this.run(state, context) ?? updateState(state, value);
+            return this.run(state, context) ?? updateState(state, value as U);
         });
     }
     /** @deprecated Use instead `.apply(manyAccum)` */